Safari-611.1.10/LayoutTests/fast/dom/HTMLInputElement/input-type-file-security-expected.txt

Tests that input.value is properly reset when changing the input type to 'file'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.type is "text"
input.value = '/etc/hosts'
PASS input.value is "/etc/hosts"
input.type = 'file'
PASS input.value is ""
PASS input.getAttribute('value') is null
PASS successfullyParsed is true

TEST COMPLETE
